+++ txnbuild/transaction.go
@@ -93,7 +93,18 @@ func (tx *Transaction) SetDefaultFee() {
 // Build for Transaction completely configures the Transaction. After calling Build,
 // the Transaction is ready to be serialised or signed.
 func (tx *Transaction) Build() error {
-	// to do: check if tx is already signed
+	// If transaction envelope has been signed, don't build transaction
+	if tx.xdrEnvelope != nil {
+		if tx.xdrEnvelope.Signatures != nil {
+			return errors.New("transaction has already been signed, so cannot be rebuilt.")
+		}
+		// clear the existing XDR so we don't append to any existing fields
+		tx.xdrEnvelope = &xdr.TransactionEnvelope{}
+		tx.xdrEnvelope.Tx = xdr.Transaction{}
+	}
+
+	// reset tx.xdrTransaction
+	tx.xdrTransaction = xdr.Transaction{}
 
 	accountID := tx.SourceAccount.GetAccountID()
 	// Public keys start with 'G'
@@ -147,10 +158,8 @@ func (tx *Transaction) Build() error {
 	tx.SetDefaultFee()
 
 	// Initialise transaction envelope
-	if tx.xdrEnvelope == nil {
-		tx.xdrEnvelope = &xdr.TransactionEnvelope{}
-		tx.xdrEnvelope.Tx = tx.xdrTransaction
-	}
+	tx.xdrEnvelope = &xdr.TransactionEnvelope{}
+	tx.xdrEnvelope.Tx = tx.xdrTransaction
 
 	return nil
 }